==> scripts/02_dea.Rmd <==
---
title: "Analysis"
output: html_document
---

```{r library}
library(tidyverse)
library(data.table)

library(lubridate)

library(gganimate)
library(ggthemes)
library(geofacet)
library(ggthemes)
```

```{r utilities}
br_states_names <- as_tibble(geofacet::br_states_grid1)

df_states_region <- tibble(code = br_states_names$code,
                           name = br_states_names$name,
                           region = c(rep("North", 4), rep("Northeast", 2),
                                      rep("North", 2), rep("Northeast", 3),
                                      "North", "Midwest", rep("Northeast", 2),
                                      rep("Midwest", 2), "Southeast", rep("Northeast", 2),
                                      "Midwest", "Southeast", "South", "Southeast",
                                      "Southeast", rep("South", 2)))
```

```{r reading}
day <- "26_04_2020"

df <- fread(paste0("./data/treated/deaths_br_", 
                   day,
                   "_by_gender_age_date_state.csv"))
```

```{r}
df <- df %>% 
  as_tibble() %>% 
  mutate(date = ymd(date))
```

# Graph 1 - Daily evolution in pyramid chart

Apparently we need to have all the data: 
https://purrr.tidyverse.org/reference/cross.html

```{r}
df_pyramid_all <- purrr::cross_df(
  .l = list(
    "state" = df %>% select(state) %>% unique() %>% pull(),
    "date" = df %>% select(date) %>% unique() %>% pull(),
    "gender" = df %>% select(gender) %>% unique() %>% pull(),
    "age_group" = df %>% select(age_group) %>% unique() %>% pull())
  )

df_pyramid_all <- df_pyramid_all %>% 
  mutate(date = lubridate::as_date(date))
```

```{r}
df_pyramid_all <- df_pyramid_all %>% 
  left_join(df, by = c("state", "date", "age_group", "gender")) %>% 
  mutate(number_deaths = ifelse(is.na(number_deaths), 0, number_deaths))
```

```{r}
age_group_levels <- df_pyramid_all %>% pull(age_group) %>% unique() %>% sort()
age_group_levels[1] <- "< 9"
age_group_levels[10] <- age_group_levels[11]
age_group_levels[11] <- "> 100"

df_pyramid_all_cumulative_deaths <- df_pyramid_all %>% 
  mutate(age_group = factor(age_group, levels = age_group_levels)) %>% 
  mutate(gender = ifelse(gender == "M", "Male", "Female")) %>% 
  arrange(state, age_group, gender, date) %>% 
  group_by(state, age_group, gender) %>% 
  mutate(number_deaths_cum = cumsum(number_deaths)) %>% 
  left_join(df_states_region, by = c("state" = "code"))
```

```{r}
g1 <- df_pyramid_all_cumulative_deaths %>% 
  mutate(number_deaths_cum = ifelse(gender == "Female", -number_deaths_cum, number_deaths_cum)) %>% 
  #filter(date == dmy("24/04/2020")) %>% 
  ggplot(aes(x = age_group, y = number_deaths_cum)) +
  geom_col(aes(fill = gender)) +
  scale_y_continuous(breaks = c(-200, -100, 0, 100, 200, 300),
                     labels = as.character(c(200, 100, 0, 100, 200, 300))) +
  labs(title = "Evolution of COVID-19 deaths (confirmed and suspected) across brazilian states - by gender and age group",
       x = "Age group", 
       y = "# of deaths",
       caption = "Source: https://transparencia.registrocivil.org.br/especial-covid\nVis: @estevaovilela with ggplot2",
       fill = "Gender") +
  coord_flip() +
  facet_wrap(~ name) +
  theme_linedraw() +
  theme(axis.text.y = element_text(size = 7),
        panel.spacing.x = unit(1, "lines"),
        panel.spacing.y = unit(0.5, "lines"),
        plot.title = element_text(face = "bold", size = 12),
        plot.caption = element_text(face = "bold", size = 9),
        legend.title = element_text(face = "bold", size = 10),
        legend.text = element_text(face = "bold", size = 9),
        axis.title.x = element_text(face = "bold", size = 10),
        axis.title.y = element_text(face = "bold", size = 10),
        strip.background = element_rect(fill = "white"),
        strip.text.x = element_text(color = "black", face = "bold", size = 9))

anim1 <- g1 +
  # Here comes the gganimate specific bits
  labs(subtitle = "Data collect in 26/04/2020. Date: {frame_time}") +
  transition_time(date) +
  theme(plot.subtitle = element_text(face = "bold", size = 9))
  #ease_aes('linear')

animate(anim1, renderer = gifski_renderer(), width = 1000, height = 800)

anim_save("./data/figures/evolution_covid19.gif")
```

# Graph 2 - Using facet_geo

```{r}
g2 <- df_pyramid_all_cumulative_deaths %>% 
  mutate(number_deaths_cum = ifelse(gender == "Female", -number_deaths_cum, number_deaths_cum)) %>% 
  filter(date == dmy("26/04/2020")) %>% 
  ggplot(aes(x = age_group, y = number_deaths_cum)) +
  geom_col(aes(fill = gender)) +
  scale_y_continuous(breaks = c(-200, -100, 0, 100, 200, 300),
                     labels = as.character(c(200, 100, 0, 100, 200, 300))) +
  labs(title = "Evolution of COVID-19 deaths (confirmed and suspected) across brazilian states - by gender and age group",
       subtitle = "Data collect in 26/04/2020",
       x = "Age group", 
       y = "# of deaths",
       caption = "Source: https://transparencia.registrocivil.org.br/especial-covid\nVis: @estevaovilela with ggplot2",
       fill = "Gender") +
  coord_flip() +
  geofacet::facet_geo(~ state, grid = "br_states_grid1", label = "name") +
  theme_linedraw() +
  theme(axis.text.y = element_text(size = 5),
        panel.spacing.x = unit(1, "lines"),
        panel.spacing.y = unit(0.5, "lines"),
        plot.title = element_text(face = "bold", size = 12),
        plot.caption = element_text(face = "bold", size = 9),
        legend.title = element_text(face = "bold", size = 10),
        legend.text = element_text(face = "bold", size = 9),
        axis.title.x = element_text(face = "bold", size = 10),
        axis.title.y = element_text(face = "bold", size = 10),
        strip.background = element_rect(fill = "white"),
        strip.text.x = element_text(color = "black", face = "bold", size = 9))

ggsave("./data/figures/geo_states_covid19.jpeg", 
       # inches
       width = 10.416, height = 8.33)
```
