==> scripts/04_dea_monthly_data_excess_of_deaths.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r library}
library(tidyverse)
library(data.table)

library(lubridate)

library(gganimate)
library(ggthemes)
library(geofacet)
```

```{r utilities}
br_states_names <- as_tibble(geofacet::br_states_grid1)

df_states_aux <- tibble(code = br_states_names$code,
                        name = br_states_names$name)
```

```{r reading}
day <- "06_05_2020"

df_states <- fread(paste0("./data/treated/all_deaths_br_", 
                         day,
                         "_by_month_state_since_2015.csv"))

df_cities <- fread(paste0("./data/treated/all_deaths_br_", 
                         day,
                         "_by_month_city_since_2015.csv"))
```

```{r}
df_states <- df_states %>% 
  as_tibble() %>% 
  left_join(df_states_aux, by = c("state" = "code"))

df_cities <- df_cities %>% 
  as_tibble() %>% 
  left_join(df_states_aux, by = c("state" = "code"))
```

# State-level

```{r}
df_states_filtered <- df_states %>% 
  filter(year != 2020)

df_states_filtered_mean <- df_states %>% 
  filter(year != 2020) %>% 
  group_by(name, month) %>% 
  summarise(monthly_avg = mean(number_deaths))
  

ggplot() +
  geom_line(data = df_states_filtered,
            aes(x = month, y = number_deaths, group = year), color = alpha("grey", 0.7)) +
  geom_line(data = df_states_filtered_mean, 
            aes(x = month, y = monthly_avg), colour = "black", linetype = "dashed") +
  geom_line(data = df_states %>% filter(year == 2020, month != 5), 
            aes(x = month, y = number_deaths, group = year), colour = alpha("red", 1)) +
  scale_x_continuous(breaks = seq(from = 1, to = 12, by = 1),
                     labels = as.character(seq(from = 1, to = 12, by = 1))) +
  labs(title = "'Excess of deaths' in brazilian states: monthly average, 2015-2019 and 2020",
       subtitle = "Data collect in 06/05/2020, Y-axis is at free scale",
       x = "Month", 
       y = "# of deaths",
       caption = "Source: https://transparencia.registrocivil.org.br/registros\nVis: @estevaovilela with ggplot2") +
  facet_wrap(~name, scales = "free_y") +
  theme_linedraw() +
  theme(axis.text.y = element_text(size = 7),
        panel.spacing.x = unit(1, "lines"),
        panel.spacing.y = unit(0.5, "lines"),
        plot.title = element_text(size = 12),
        plot.caption = element_text(face = "bold", size = 9),
        axis.title.x = element_text(face = "bold", size = 10),
        axis.title.y = element_text(face = "bold", size = 10),
        strip.background = element_rect(fill = "white"),
        strip.text.x = element_text(color = "black", face = "bold", size = 9))

ggsave("./data/figures/excess_of_deaths_states.jpeg", 
       # inches
       width = 10.416, height = 8.33)
```

# Capital

```{r capitais}
# Maceio, São Luis, Cuiaba, Macapa, Boa Vista, Florianopolis, Joao Pessoa
df_capitals <- tibble(capital = c(
  "Porto Alegre", "Florianopolis", "Curitiba",
  "São Paulo", "Rio de Janeiro", "Belo Horizonte",
  "Vitória", "Brasília", "Goiânia", "Campo Grande",
  "Cuiaba", "Salvador", "Maceio", "Aracaju", "Recife",
  "Joao Pessoa", "Natal", "Fortaleza", "São Luis",
  "Teresina", "Manaus", "Belém", "Rio Branco", 
  "Porto Velho", "Macapa", "Palmas", "Boa Vista"))
```

```{r}
df_cities_filtered <- df_cities %>% 
  semi_join(df_capitals, by = c("city" = "capital")) %>%
  filter(year != 2020)

df_cities_filtered_mean <- df_cities %>% 
  semi_join(df_capitals, by = c("city" = "capital")) %>% 
  filter(year != 2020) %>% 
  group_by(city, month) %>% 
  summarise(monthly_avg = mean(number_deaths))

ggplot() +
  geom_line(data = df_cities_filtered,
            aes(x = month, y = number_deaths, group = year), color = alpha("grey", 0.7)) +
  geom_line(data = df_cities_filtered_mean, 
            aes(x = month, y = monthly_avg), colour = "black", linetype = "dashed") +
  geom_line(data = df_cities %>% semi_join(df_capitals, by = c("city" = "capital")) %>% 
              filter(year == 2020), 
            aes(x = month, y = number_deaths, group = year), colour = alpha("red", 1)) +
  scale_x_continuous(breaks = seq(from = 1, to = 12, by = 1),
                     labels = as.character(seq(from = 1, to = 12, by = 1))) +
  labs(title = "'Excess of deaths' in brazilian capitals: monthly average, 2015-2019 and 2020",
       subtitle = "Data collect in 29/04/2020, Y-axis is at free scale",
       x = "Month", 
       y = "# of deaths",
       caption = "Source: https://transparencia.registrocivil.org.br/registros\nVis: @estevaovilela with ggplot2") +
  facet_wrap(~city, scales = "free_y") +
  theme_linedraw() +
  theme(axis.text.y = element_text(size = 7),
        panel.spacing.x = unit(1, "lines"),
        panel.spacing.y = unit(0.5, "lines"),
        plot.title = element_text(size = 12),
        plot.caption = element_text(face = "bold", size = 9),
        axis.title.x = element_text(face = "bold", size = 10),
        axis.title.y = element_text(face = "bold", size = 10),
        strip.background = element_rect(fill = "white"),
        strip.text.x = element_text(color = "black", face = "bold", size = 9))
```

# Norte e Nordeste

```{r capitais}
# Maceio, São Luis, Cuiaba, Macapa, Boa Vista, Florianopolis, Joao Pessoa
df_capitals <- tibble(capital = c(
  "Porto Alegre", "Florianopolis", "Curitiba",
  "São Paulo", "Rio de Janeiro", "Belo Horizonte",
  "Vitória", "Brasília", "Goiânia", "Campo Grande",
  "Cuiaba", "Salvador", "Maceio", "Aracaju", "Recife",
  "Joao Pessoa", "Natal", "Fortaleza", "São Luís",
  "Teresina", "Manaus", "Belém", "Rio Branco", 
  "Porto Velho", "Macapa", "Palmas", "Boa Vista"),
  state = c("RS", "SC", "PR", "SP", "RJ", "MG",
            "ES", "DF", "GO", "MS", "MT", "BA", 
            "AL", "SE", "PE", "PB", "RN", "CE", 
            "MA", "PI", "AM", "PA", "AC", "RO",
            "AP", "TO", "RR"))
```

```{r two df}
df_ne_no_19 <- df_cities %>% 
  filter(state %in% c("AC", "AP", "AM", "PA", "MA", "CE", "TO", "PI",
                      "RR", "PB", "RN", "RO", "PE", "AL", "BA", "SE")) %>% 
  semi_join(df_capitals, by = c("city" = "capital", "state")) %>%
  filter(year == 2019)

df_ne_no_20 <- df_cities %>% 
  filter(state %in% c("AC", "AP", "AM", "PA", "MA", "CE", "TO", "PI",
                      "RR", "PB", "RN", "RO", "PE", "AL", "BA", "SE")) %>% 
  semi_join(df_capitals, by = c("city" = "capital", "state")) %>%
  filter(year == 2020, month != 5)

ggplot() +
  geom_line(data = df_ne_no_19,
            aes(x = month, y = number_deaths, group = year), color = alpha("black", 1)) +
  geom_line(data = df_ne_no_20, 
            aes(x = month, y = number_deaths, group = year), colour = alpha("red", 1)) +
  scale_x_continuous(breaks = seq(from = 1, to = 12, by = 1),
                     labels = as.character(seq(from = 1, to = 12, by = 1))) +
  labs(title = "'Excesso de mortalidade geral' nas capitais do Norte e Nordeste - 2019/2020",
       subtitle = "Dados coletados em 06/05/2020, Eixo-Y em escala livre",
       x = "Mês", 
       y = "# de óbitos registrados",
       caption = "Fonte: https://transparencia.registrocivil.org.br/registros") +
  facet_wrap(~city, scales = "free_y") +
  theme_linedraw() +
  theme(axis.text.y = element_text(size = 7),
        panel.spacing.x = unit(1, "lines"),
        panel.spacing.y = unit(0.5, "lines"),
        plot.title = element_text(size = 12),
        plot.caption = element_text(face = "bold", size = 9),
        axis.title.x = element_text(face = "bold", size = 10),
        axis.title.y = element_text(face = "bold", size = 10),
        strip.background = element_rect(fill = "white"),
        strip.text.x = element_text(color = "black", face = "bold", size = 9))
```

```{r}
df_ne_no <- df_cities %>% 
  filter(state %in% c("AC", "AP", "AM", "PA", "MA", "CE", "TO", "PI",
                      "RR", "PB", "RN", "RO", "PE", "AL", "BA", "SE")) %>% 
  semi_join(df_capitals, by = c("city" = "capital", "state")) %>%
  filter((year == 2020 & month != 5) | year == 2019)

ggplot() +
  geom_line(data = df_ne_no,
            aes(x = month, y = number_deaths, group = year, color = as.factor(year))) +
  scale_color_manual(values = c("black", "red")) +
  scale_x_continuous(breaks = seq(from = 1, to = 12, by = 1),
                     labels = as.character(seq(from = 1, to = 12, by = 1))) +
  labs(title = "'Excesso de mortalidade geral' nas capitais do Norte e Nordeste - 2019/2020",
       subtitle = "Dados coletados em 06/05/2020, Eixo-Y em escala livre",
       x = "Mês", 
       y = "# de óbitos registrados",
       caption = "Fonte: https://transparencia.registrocivil.org.br/registros",
       color = "Ano") +
  facet_wrap(~city, scales = "free_y") +
  theme_linedraw() +
  theme(axis.text.y = element_text(size = 7),
        panel.spacing.x = unit(1, "lines"),
        panel.spacing.y = unit(0.5, "lines"),
        plot.title = element_text(size = 12),
        plot.caption = element_text(face = "bold", size = 9),
        axis.title.x = element_text(face = "bold", size = 10),
        axis.title.y = element_text(face = "bold", size = 10),
        strip.background = element_rect(fill = "white"),
        strip.text.x = element_text(color = "black", face = "bold", size = 9))

# salvado manualmente
# ggsave("./data/figures/excess_of_deaths_north_northeast_states.jpeg", 
#        # inches, 1000, 
#        width = 10.416, height = 8.33)
```

```{r loess}
capitais_nordeste <- df_ne_no %>% 
  pull(city) %>% 
  unique() %>% 
  sort()

loess_capitals <- vector("numeric")

for(i in 1:length(capitais_nordeste)) {
    loess_capitals_temp <- predict(loess(data = df_ne_no[df_ne_no$year == 2019,], 
                                         formula = number_deaths ~ month, 
                                         subset = city == capitais_nordeste[i], 
                                         span = 0.5))
    
    loess_capitals <- append(x = loess_capitals, values = loess_capitals_temp)
}

loess_capitals <- append(x = loess_capitals, values = c(rep(NA, 64)))

df_ne_no$number_deaths_loess <- vector("numeric", length = length(df_ne_no$city))

df_ne_no_smoothed <- df_ne_no %>% 
  arrange(year, city) %>% 
  mutate(number_deaths_loess = loess_capitals) %>% 
  mutate(number_deaths_loess = ifelse(is.na(number_deaths_loess), 
                                      number_deaths,
                                      number_deaths_loess))

ggplot() +
  geom_line(data = df_ne_no_smoothed,
            aes(x = month, y = number_deaths_loess, group = year, color = as.factor(year))) +
  scale_color_manual(values = c("black", "red")) +
  scale_x_continuous(breaks = seq(from = 1, to = 12, by = 1),
                     labels = as.character(seq(from = 1, to = 12, by = 1))) +
  labs(title = "'Excesso de mortalidade geral' nas capitais do Norte e Nordeste - 2019/2020",
       subtitle = "Dados coletados em 06/05/2020, Eixo-Y em escala livre",
       x = "Mês", 
       y = "# de óbitos registrados",
       caption = "Fonte: https://transparencia.registrocivil.org.br/registros",
       color = "Ano") +
  facet_wrap(~city, scales = "free_y") +
  theme_linedraw() +
  theme(axis.text.y = element_text(size = 7),
        panel.spacing.x = unit(1, "lines"),
        panel.spacing.y = unit(0.5, "lines"),
        plot.title = element_text(size = 12),
        plot.caption = element_text(face = "bold", size = 9),
        axis.title.x = element_text(face = "bold", size = 10),
        axis.title.y = element_text(face = "bold", size = 10),
        strip.background = element_rect(fill = "white"),
        strip.text.x = element_text(color = "black", face = "bold", size = 9))

```
