==> scripts/04.1_dea_monthly_data_excess_of_deaths_metrop.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r library}
library(tidyverse)
library(data.table)

library(lubridate)

library(ggthemes)
```

```{r reading}
day <- "05_08_2020"

df_brazil <- fread(paste0("./data/treated/all_deaths_br_", 
                          day, 
                          "_by_month_metropolitan_region_since_2019_ibge_code.csv")) %>% 
  as_tibble()

df_ecuador <- fread("./data/treated/all_deaths_ecuador_by_month_provincia_since_2018.csv") %>% 
  as_tibble()
```

```{r}
tema <- theme(axis.text.y = element_text(face = "bold", size = 9),
              axis.text.x = element_text(face = "bold", size = 9),
              panel.spacing.x = unit(1, "lines"),
              panel.spacing.y = unit(0.5, "lines"),
              plot.title = element_text(size = 12),
              plot.caption = element_text(face = "bold", size = 9),
              axis.title.x = element_text(size = 10),
              axis.title.y = element_text(size = 10),
              strip.background = element_rect(fill = "white"),
              strip.text.x = element_text(color = "black", face = "bold", size = 9),
              legend.title = element_blank())
```

# Brazil
## Whole year

```{r}
df_br_january_june <- df_brazil %>% 
  # some cities doesn't have any deaths recorded
  filter(!is.na(year)) %>% 
  filter(month <= 6) %>% 
  group_by(rm, capital, year) %>% 
  summarise(total = sum(number_deaths)) %>% 
  spread(year, total) %>% 
  mutate(dif = (`2020` - `2019`) / `2019` * 100)

df_br_january_june <- df_br_january_june %>% 
  ungroup() %>% 
  mutate(capital = ifelse(capital == 1, "Capital of Metropolitan Area", "Metropolitan Area without Capital"))

ggplot() +
  geom_segment(data = df_br_january_june %>% select(rm, capital, dif) %>% spread(capital, dif) %>% arrange(rm),
               aes(x = rm, xend = rm, 
                   y = `Capital of Metropolitan Area`, yend = `Metropolitan Area without Capital`), 
               color = alpha("black", 0.7)) +
  geom_point(data = df_br_january_june %>% arrange(rm),
            aes(x = rm, y = dif, color = capital)) +
  coord_flip() +
  scale_y_continuous(breaks = seq(from = -50, to = 80, by = 10),
                     labels = as.character(seq(from = -50, to = 80, by = 10))) +
  labs(title = "Figure 1: Difference, in percent, between the number of deaths registered in 2020 compared to 2019 in the months January to June.\nBrazilian Metropolitan Capitals and Metropolitan Areas without Capital.",
       subtitle = "Data collected on August, 5",
       x = "Metropolitan Area", 
       y = "Difference, in percent (%)",
       caption = "Source: https://transparencia.registrocivil.org.br/registros") +
  theme_linedraw() +
  tema
```

## Temporal

```{r}
df_br_temporal <- df_brazil %>% 
   # some cities doesn't have any deaths recorded
  filter(!is.na(year)) %>% 
  group_by(rm, year, month, capital) %>% 
  summarise(number_deaths = sum(number_deaths))

df_br_temporal <- df_br_temporal %>% 
  ungroup() %>% 
  mutate(capital = ifelse(capital == 1, "Capital of Metropolitan Area", "Metropolitan Area without Capital")) %>% 
  filter((year == 2019 & month <= 6) | (year == 2020 & month <= 5))

ggplot() +
  geom_line(data = df_br_temporal %>% filter(capital == "Capital of Metropolitan Area"),
            aes(x = month, y = number_deaths, group = year, color = factor(year))) +
  geom_line(data = df_br_temporal %>% filter(capital != "Capital of Metropolitan Area"),
            aes(x = month, y = number_deaths, group = year, color = factor(year)),
            linetype = "dashed") +
  scale_color_manual(values = c("black", "red")) +
  facet_wrap(~ rm, scales = "free_y") +
  scale_x_continuous(breaks = seq(from = 1, to = 12, by = 1),
                     labels = as.character(seq(from = 1, to = 12, by = 1))) +
  labs(title = "Figure 2: Temporal evolution of excess mortality across main metropolitan areas in Brazil, monthly comparison between 2019-2020",
       subtitle = "Data collected on June, 6; Solid line: Capital of Metropolitan Area, Dashed line: Metropolitan Area without Capital",
       x = "Month", 
       y = "# of deaths",
       color = "Year",
       caption = "Source: https://transparencia.registrocivil.org.br/registros") +
  theme_linedraw() +
  tema
```

# Ecuador

## Temporal

```{r}
ggplot() +
  geom_line(data = df_ecuador,
            aes(x = month, y = number_deaths, group = year, color = factor(year))) +
  scale_color_manual(values = c("black", "red", "blue")) +
  facet_wrap(~ provincia, scales = "free_y") +
  scale_x_continuous(breaks = seq(from = 1, to = 12, by = 1),
                     labels = as.character(seq(from = 1, to = 12, by = 1))) +
  labs(title = "Figure 2: Temporal evolution of excess mortality across provincias in Ecuador, monthly comparison between 2018-2020",
       subtitle = "Data collected on May, 23",
       x = "Month", 
       y = "# of deaths",
       color = "Year",
       caption = "Source: https://www.registrocivil.gob.ec/cifras/") +
  theme_linedraw() +
  tema
```
